+++ api-controllers/ServiceFabrikAdminController.js
@@ -16,7 +16,7 @@ const bosh = require('../data-access-layer/bosh');
 const ScheduleManager = require('../broker/lib/jobs');
 const BackupReportManager = require('../reports');
 const CONST = require('../common/constants');
-const maintenanceManager = require('../broker/lib/maintenance').maintenanceManager;
+const maintenanceManager = require('../maintenance').maintenanceManager;
 const Conflict = errors.Conflict;
 const Forbidden = errors.Forbidden;
 const BadRequest = errors.BadRequest;
