+++ broker/JobScheduler.js
@@ -11,7 +11,7 @@ const logger = require('../common/logger');
 const config = require('../common/config');
 const utils = require('./lib/utils');
 const errors = require('../common/errors');
-const maintenanceManager = require('./lib/maintenance').maintenanceManager;
+const maintenanceManager = require('../maintenance').maintenanceManager;
 const serviceFabrikClient = require('../data-access-layer/cf').serviceFabrikClient;
 require('./lib/fabrik');
