+++ broker/lib/jobs/Scheduler.js
@@ -20,7 +20,7 @@ const errors = require('../../../common/errors');
 const BadRequest = errors.BadRequest;
 const ServiceUnavailable = errors.ServiceUnavailable;
 const ServiceInMaintenance = errors.ServiceInMaintenance;
-const maintenanceManager = require('../maintenance').maintenanceManager;
+const maintenanceManager = require('../../../maintenance').maintenanceManager;
 const MONGO_TO_BE_INITIALIZED = 0;
 const MONGO_INIT_FAILED = 1;
 const MONGO_INIT_SUCCEEDED = 2;
