+++ maintenance/MaintenanceManager.js
@@ -1,11 +1,11 @@
 const _ = require('lodash');
 const Promise = require('bluebird');
-const utils = require('../utils');
-const Repository = require('../../../common/db').Repository;
-const logger = require('../../../common/logger');
-const CONST = require('../../../common/constants');
-const errors = require('../../../common/errors');
-const config = require('../../../common/config');
+const utils = require('../common/utils');
+const Repository = require('..//common/db').Repository;
+const logger = require('../common/logger');
+const CONST = require('../common/constants');
+const errors = require('../common/errors');
+const config = require('../common/config');
 
 class MaintenanceManager {
