+++ test/test_broker/JobScheduler.spec.js
@@ -6,7 +6,7 @@ const moment = require('moment');
 const CONST = require('../../common/constants');
 const config = require('../../common/config');
 const errors = require('../../common/errors');
-const maintenanceManager = require('../../broker/lib/maintenance').maintenanceManager;
+const maintenanceManager = require('../../maintenance').maintenanceManager;
 const serviceFabrikClient = require('../../data-access-layer/cf').serviceFabrikClient;
 const logger = require('../../common/logger');
